requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
gunicorn==22.0.0
django-cors-headers
